refactor: move command files into src/commands/ as *.command.ts
Consistent naming across commands: add.command.ts, sync.command.ts,
init.command.ts, jsr.command.ts, all under src/commands/. Only import
paths changed (adjusted for the extra directory level) — no behavior
change.
