+++ src/commands/sync.command.ts
@@ -1,8 +1,8 @@
 import * as p from '@clack/prompts';
 import pc from 'picocolors';
-import { findWorkspaceRoot, discoverPackages, detectRepo, type Pkg } from './workspace.js';
-import { npmAuthCheck, checkNpmVersion, listTrust, configureTrust, revokeTrust, warmNpmAuth, publishedNames } from './npm.js';
-import { npmArgs, selectorsOf, type Ctx } from './args.js';
+import { findWorkspaceRoot, discoverPackages, detectRepo, type Pkg } from '../workspace.js';
+import { npmAuthCheck, checkNpmVersion, listTrust, configureTrust, revokeTrust, warmNpmAuth, publishedNames } from '../npm.js';
+import { npmArgs, selectorsOf, type Ctx } from '../args.js';
 import {
   resolveTargets,
   validateTrustSettings,
@@ -12,9 +12,9 @@ import {
   describeTrustDiff,
   describeConfig,
   applyIgnore,
-} from './core.js';
-import { loadConfig } from './config.js';
-import { hatchSpinner, hatchIntro, otpBoxReminder, reportNpmAuth, note } from './ui.js';
+} from '../core.js';
+import { loadConfig } from '../config.js';
+import { hatchSpinner, hatchIntro, otpBoxReminder, reportNpmAuth, note } from '../ui.js';
 
 /**
  * `fledgling sync` — reconcile trusted publishing across the workspace.
